代码更新

+++ src/main/java/com/niule/znxj/web/dao/ExceptionhandlerinfoMapper.java
@@ -10,7 +10,7 @@ public interface ExceptionhandlerinfoMapper {
 
     int deleteByExample(ExceptionhandlerinfoExample example);
 
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(Exceptionhandlerinfo record);
 
@@ -18,7 +18,7 @@ public interface ExceptionhandlerinfoMapper {
 
     List<Exceptionhandlerinfo> selectByExample(ExceptionhandlerinfoExample example);
 
-    Exceptionhandlerinfo selectByPrimaryKey(Integer id);
+    Exceptionhandlerinfo selectByPrimaryKey(Long id);
 
     int updateByExampleSelective(@Param("record") Exceptionhandlerinfo record, @Param("example") ExceptionhandlerinfoExample example);
